[#98] Allow execution of delete command with k8s

+++ cmd/deletion/delete-topic.go
@@ -2,6 +2,7 @@ package deletion
 
 import (
 	"github.com/deviceinsight/kafkactl/operations"
+	"github.com/deviceinsight/kafkactl/operations/k8s"
 	"github.com/deviceinsight/kafkactl/output"
 	"github.com/spf13/cobra"
 )
@@ -13,8 +14,10 @@ func newDeleteTopicCmd() *cobra.Command {
 		Short: "delete a topic",
 		Args:  cobra.MinimumNArgs(1),
 		Run: func(cmd *cobra.Command, args []string) {
-			if err := (&operations.TopicOperation{}).DeleteTopics(args); err != nil {
-				output.Fail(err)
+			if !(&k8s.K8sOperation{}).TryRun(cmd, args) {
+				if err := (&operations.TopicOperation{}).DeleteTopics(args); err != nil {
+					output.Fail(err)
+				}
 			}
 		},
 		ValidArgsFunction: operations.CompleteTopicNames,
